requirements.txt
asgiref==3.8.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dj-database-url==3.0.0
Django==5.2.1
django-appconf==1.1.0
django-azure-auth==2.3.0
django-environ==0.12.0
django-select2==8.4.1
et_xmlfile==2.0.0
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
msal==1.33.0
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2

requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
xlwings==0.33.14
